requirements.txt
# Core functionality
sentence-transformers==2.5.1
qdrant-client==1.6.3
pymongo==4.6.3
requests==2.31.0
tqdm==4.66.4
opencv-python 
decord
# Gradio app
gradio==4.26.0

# Video processing
ffmpeg-python==0.2.0

# Optional/used during earlier stages
spacy==3.7.2
webvtt-py==0.4.6

# General utilities
numpy==1.26.4
scipy==1.13.0

# For BERTopic if used anywhere in extension
bertopic==0.17.0
umap-learn==0.5.7
hdbscan==0.8.40

# Transformers for Gemma / Ollama fallback
transformers==4.51.3
huggingface-hub==0.30.2
tokenizers==0.21.1
